Extra fields for SSHKey and use QuerySet instead of Manager
