+++ readthedocs/projects/querysets.py
@@ -1,3 +1,4 @@
+# -*- coding: utf-8 -*-
 """Project model QuerySet classes"""
 
 from __future__ import absolute_import
@@ -6,8 +7,10 @@ from django.db import models
 from django.db.models import Q
 from guardian.shortcuts import get_objects_for_user
 
-from . import constants
 from readthedocs.core.utils.extend import SettingsOverrideObject
+
+from . import constants
+from .ssh import generate_public_from_private_key
 
 
 class ProjectQuerySetBase(models.QuerySet):
@@ -159,3 +162,23 @@ class FeatureQuerySet(models.QuerySet):
             Q(projects=project) |
             Q(default_true=True, add_date__gt=project.pub_date)
         ).distinct()
+
+
+class SSHKeyQuerySet(models.QuerySet):
+
+    def create_from_string(self, private_key, project):
+        """
+        Create a SSHKey instance from a private key.
+
+        :param private_key: private key string
+        :param project: project to attach the SSH key
+        """
+        key = self.create(
+            private_key=private_key,
+            public_key=generate_public_from_private_key(
+                private_key,
+                text_format=True,
+            ),
+            project=project,
+        )
+        return key
